remove prototype list attachment commit

+++ application/quality/model/ProcedureListModel.php
@@ -46,6 +46,13 @@ class ProcedureListModel extends Model
     public function delProcedureList($id)
     {
         try{
+            // 关联删除think_procedure_list_sublist
+            $proceSub = new ProcedureListSublistModel();
+            $bol = $proceSub->delProceSubByGroupId($id);
+            if($bol['code'] == 0){
+                return $bol;
+            }
+
             $this->where('id', $id)->delete();
             return ['code' => 1, 'data' => '', 'msg' => '删除成功'];
 
@@ -57,5 +64,18 @@ class ProcedureListModel extends Model
     public function getOne($id)
     {
         return $this->where('id', $id)->find();
+    }
+
+    public function delProceListByGroupId($groupId){
+        $idArr = $this->where('group_id',$groupId)->column('id');
+        if(count($idArr) > 0){
+            foreach ($idArr as $k=>$v){
+                $bol = $this->delProcedureList($v);
+                if($bol['code'] == 0){
+                    return $bol;
+                }
+            }
+        }
+        return ['code' => 1, 'data' => '', 'msg' => 'List删除成功'];
     }
 }
